+++ Graphics/VulkanGraphicsEngine/VulkanGraphicsEngineSwapChain.cpp
@@ -27,7 +27,7 @@ VkExtent2D VulkanGraphicsEngine::chooseSwapExtent(const VkSurfaceCapabilitiesKHR
 	if (capabilities.currentExtent.width != std::numeric_limits<uint32_t>::max())
 		return capabilities.currentExtent;
 	else {
-		VkExtent2D actualExtent = {width, height};
+		VkExtent2D actualExtent = {(uint32_t) width, (uint32_t) height};
 		actualExtent.width = std::max(capabilities.minImageExtent.width, std::min(capabilities.maxImageExtent.width, actualExtent.width));
 		actualExtent.height = std::max(capabilities.minImageExtent.height, std::min(capabilities.maxImageExtent.height, actualExtent.height));
 		return actualExtent;
